tests/wpt/web-platform-tests/offscreen-canvas/the-offscreen-canvas/offscreencanvas.transfercontrol.to.offscreen.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>
<link rel="help" href="https://html.spec.whatwg.org/#dom-canvas-transfercontroltooffscreen">
<script>

test(function() {
    var placeholder = document.createElement('canvas');
    placeholder.width = 100;
    placeholder.height = 50;
    var offscreenCanvas = placeholder.transferControlToOffscreen();
    assert_equals(offscreenCanvas.width, 100);
    assert_equals(offscreenCanvas.height, 50);
}, "Test that an OffscreenCanvas generated by transferControlToOffscreen gets correct width and height");

test(function() {
    var placeholder = document.createElement('canvas');
    placeholder.width = 100;
    placeholder.height = 50;
    var offscreenCanvas = placeholder.transferControlToOffscreen();
    assert_throws("InvalidStateError", function() { placeholder.getContext('2d'); });
}, "Test that calling getContext on a placeholder canvas that has already transferred its control throws an exception");

test(function() {
    var placeholder = document.createElement('canvas');
    placeholder.width = 100;
    placeholder.height = 50;
    var offscreenCanvas = placeholder.transferControlToOffscreen();
    assert_throws("InvalidStateError", function() { placeholder.transferControlToOffscreen(); });
}, "Test that calling transferControlToOffscreen twice throws an exception");

async_test(function(t) {
    var placeholder = document.createElement('canvas');
    placeholder.width = 10;
    placeholder.height = 20;
    var offscreenCanvas = placeholder.transferControlToOffscreen();
    var ctx = offscreenCanvas.getContext('2d');
    t.step(function() {
        offscreenCanvas.width = 30;
        offscreenCanvas.height = 40;
        ctx.fillStyle = '#0f0';
        ctx.fillRect(0, 0, 30, 40);
        assert_equals(offscreenCanvas.width, 30);
        assert_equals(offscreenCanvas.height, 40);
        var image = offscreenCanvas.transferToImageBitmap();
        assert_equals(image.width, 30);
        assert_equals(image.height, 40);
    });
    t.step(function() {
        // Setting the size of an OffscreenCanvas does not directly update the size of its placeholder canvas.
        assert_equals(placeholder.width, 10);
        assert_equals(placeholder.height, 20);
    });
    ctx.commit();
    t.step(function() {
        // commit() doesnt't synchronously update the size of its placeholder canvas
        assert_equals(placeholder.width, 10);
        assert_equals(placeholder.height, 20);
    });
    // Set timeout acts as a sync barrier to allow commit to propagate
    setTimeout(function() {
        t.step(function() {
            // commit() asynchronously updates the size of its placeholder canvas
            assert_equals(placeholder.width, 30);
            assert_equals(placeholder.height, 40);
        });
        createImageBitmap(placeholder).then(image => {
            t.step(function() {
                assert_equals(image.width, 30);
                assert_equals(image.height, 40);
                var canvas = document.createElement('canvas');
                canvas.width = canvas.height = 50;
                var context = canvas.getContext('2d');
                context.drawImage(image, 0, 0);
                _assertPixel(canvas, 15,20, 0,255,0,255, "15,20", "0,255,0,255");
                t.done();
            });
        });
    }, 0);
}, "Verify that resizing an OffscreenCanvas with a 2d context propagates the new size to its placeholder canvas asynchronously, upon commit.");

</script>
