third_party/WebKit/LayoutTests/external/wpt/offscreen-canvas/the-offscreen-canvas/offscreencanvas.transfercontrol.to.offscreen-expected.txt

This is a testharness.js-based test.
PASS Test that an OffscreenCanvas generated by transferControlToOffscreen gets correct width and height 
PASS Test that calling getContext on a placeholder canvas that has already transferred its control throws an exception 
PASS Test that calling transferControlToOffscreen twice throws an exception 
FAIL Verify that resizing an OffscreenCanvas with a 2d context propagates the new size to its placeholder canvas asynchronously, upon commit. assert_equals: expected 30 but got 10
Harness: the test ran to completion.
